refactor(client-3d): extract Room props to scene/props/office/
Phase 3A (partial): split Room.tsx 958→411 LOC. Moved Laptop, Sofa,
PottedPlant, WaterStation, OldComputerDesk to props/office/Furniture.tsx
and VideoDisplay, Door, PictureFrame to props/office/WallObjects.tsx.
DJBooth/DJSlotOrb and BobbingGroup stay in Room.tsx since they're
tightly coupled to room layout + trampoline ripple system.
